# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Montor-ing API",
        "description": "Dokumentasi API untuk aplikasi Montor-ing (Inventory & Sales Monitoring).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Server API Utama"
        }
    ],
    "paths": {
...
        "/server/check": {
            "post": {
                "tags": [
                    "Dynamic Connection"
                ],
                "summary": "Cek Server",
                "description": "Memeriksa apakah server terdaftar dan memiliki database aktif.",
                "operationId": "ff6afa4e360de20aaf3313be1f599d1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Server valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "server_name": {
                                                    "type": "string",
                                                    "example": "Server 1"
                                                },
                                                "ip_address": {
                                                    "type": "string",
                                                    "example": "[IP_ADDRESS]"
                                                },
                                                "total_db": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "databases": {
                                                    "properties": {
                                                        "db1": {
                                                            "type": "string",
                                                            "example": "Database 1"
                                                        },
                                                        "db2": {
                                                            "type": "string",
                                                            "example": "Database 2"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Server tidak ditemukan atau tidak aktif"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari endpoint /login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}